+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Method on lamp to clear spectrum but preserve wavelength/guv_type
 - place_at_index and place_lamp_at_index API in lamp_placement module
 - Major occlusion/shadowing module
+- Add Object class
 
 ### Fixed